+++ cloud-wallet-openid4vc/src/issuance/notification.rs
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::errors::{Error, ErrorKind};
 
-use super::utils::helper::is_allowed_ascii_byte;
+use super::utils::allowed_chars::is_allowed_ascii_byte;
 
 /// Notification event types defined in OpenID4VCI §11.1.
 ///
